Implement Join Projects Functionality (#80)
* Add join button functionality

- Added "Join Project" button to the project form component.
- Implemented logic in the component to handle join project requests.
- Created `joinProject` method in `ProjectService` to send join requests to the backend.

This sub-feature allows users to request to join projects they are interested in.

* Progress

* Resolve DetachedInstanceError in delete tests for
Comment and Discussion services

- Ensure related data is accessed and response objects are created
before deleting entities to prevent DetachedInstanceError.
- Previously, there was an error in one of the tests in
discussion_test.py where a field of one of the demo-data discussions was
changed, and that change persisted in the database after the test was
completed, this has been resolved.
- Test coverage is now at 100% for the entire application.
- Tests for join_requests are next on the to-do list.

This commit addresses the issue where accessing lazy-loaded
relationships after entity deletion caused DetachedInstanceError.
By ensuring that related data is accessed and response objects are
created before deleting entities, the tests for deleting discussions
and comments now pass consistently.

* Prevent duplicate join requests and add
corresponding tests

- Added 100% test coverage for join_requests
- Updated `create_join_request` method in
`JoinRequestService` to check for existing join requests by the same
user for the same project before creating a new one. This ensures that
users cannot submit duplicate join requests.
- Added a new exception `JoinRequestAlreadyMadeException` to handle
cases where a duplicate join request is detected.
- Modified the test data setup in `join_request_data.py` to correctly
insert fake data for join requests and ensure the necessary user and
project entities exist.
- Added a test case `test_create_join_request_exists` to verify that a
duplicate join request raises the `JoinRequestAlreadyMadeException`.
- Verified that the test data is correctly inserted by printing out the
inserted join requests for debugging purposes.

This commit ensures the system prevents users from making duplicate join requests and includes the necessary tests to verify this behavior.

* Update frontend project service and backend delete
join request methods

- Updated frontend project service to take user_id and project_id when
deleting project join requests.
- Added leave project method to project service.
- Updated delete_join_request method in JoinRequestService to take
user_id and project_id as parameters.
- Modified API endpoint to match the new delete_join_request method
signature.
- Updated test cases to pass user_id and project_id when calling
delete_join_request.

This change ensures that join requests are uniquely identified by both user_id and project_id, allowing more precise deletions.

* Adjust project form view based on user roles and
edit state

- Display different views for leaders and non-leaders:
  - Show project details for non-leaders and leaders who are not editing.
  - Include join/leave buttons for non-leaders.
  - Add an edit button for leaders in the project details view.
- Display project form for leaders only in edit mode or when creating a
new project.
- Introduce `isEditing` state to toggle between edit and view modes for
leaders.
- Update HTML template to conditionally render components based on user
role and `isEditing` state.
- Update TypeScript component to handle edit mode toggling and form
submission logic.

This change ensures a clear distinction between views for different user
roles and adds edit functionality for project leaders.

* Implement join request management and
project request viewing

- Added status field to JoinRequestEntity to track 'pending', 'approved'
, and 'denied' statuses.
- Updated JoinRequestEntity to include from_model method and updated
to_join_request_response method.
- Modified JoinRequestService to handle approval and rejection of
join requests by updating the status field.
- Created new endpoints in FastAPI for approving and rejecting join
requests.
- Updated get_join_request_by_project method to return a list of
JoinRequestResponse objects.
- Implemented ProjectRequestsComponent in Angular to view join requests
for a project.
- Added methods in ProjectService to call the new API endpoints for
approving and rejecting join requests.
- Adjusted ProjectRequestsComponent to handle the responses from the
API and update the join request status in the UI.
- Added route in Angular for the new project requests page.
- Ensured proper initialization and handling in ProjectRequestsComponent
with OnInit lifecycle hook.
- Verified and tested backend API responses with cURL to ensure correct
JSON responses.

This commit enhances the functionality to manage and view project
join requests, improving the user experience for project leaders and
members.

* Enhance JWT error handling and add authentication
to APIs

- Updated `get_current_user` to raise `HTTPException` with 401 status
for expired or invalid tokens
- Added detailed error messages for better debugging and user feedback
- Ensured user authentication handles token expiry and validity checks
robustly
- Integrated authentication checks into APIs to ensure
secure access
- Modified project creation, update, and deletion to validate the
current user's identity

This commit improves the security and reliability of the authentication
flow by ensuring that expired or invalid tokens are properly handled and
appropriate error messages are returned. Additionally, it secures
project-related operations by incorporating user authentication checks.

* Simplify authentication headers management

- Implement HTTP interceptor to automatically attach authentication
headers to all HTTP requests.
- Refactor ProjectService to remove manual header attachment in each
service method.
- Ensure all API calls include the authentication token by default via
the interceptor.

This change simplifies the codebase by centralizing the authentication
logic and reducing redundancy.

* Integrate token validation interceptor across all
backend APIs

- Refactored all API endpoints to include token validation using
`get_current_user` dependency.
- Applied authorization checks to all relevant API endpoints to ensure
authenticated access.
- Improved error handling and logging for better traceability and
debugging.
- Enhanced security and integrity of operations by enforcing consistent
token validation across all services.

This comprehensive refactor aligns the backend with the frontend's
token-based authentication interceptor, ensuring secure and consistent
access control.

* Fix routing issue when deleting comments/nested
comments

- Fixed the issue where comments were not being removed from the page
without a refresh after being deleted
- Added routing for when a comment is deleted, if the comment was
attached to a project it will reroute to the project, same for discussion
and parent comments

* Enhance asthetics of project page and fix issue
with editor not popping up

* First commit in a long time

---------

+++ backend/services/project.py
@@ -1,13 +1,12 @@
 import logging
-from typing import Dict, List
-from pytest import Session
-from sqlalchemy.exc import IntegrityError
+from typing import List, Dict
+from sqlalchemy.orm import Session
+from sqlalchemy.exc import NoResultFound
 from backend.entities.project_entity import ProjectEntity
 from backend.entities.user_entity import UserEntity
-from backend.models.project import Project, ProjectCreate, ProjectResponse, ProjectUpdate
-from backend.models.user import User, UserResponse
-from backend.services.exceptions import ProjectNotFoundException
-from sqlalchemy.orm.exc import NoResultFound 
+from backend.models.project import ProjectCreate, ProjectResponse, ProjectUpdate
+from backend.services.exceptions import ProjectNotFoundException, UnauthorizedException
+from backend.models.user import UserResponse
 
 logger = logging.getLogger(__name__)
 
@@ -16,15 +15,8 @@ class ProjectService:
         self.db = db
 
     def create_project(self, project_data: ProjectCreate) -> ProjectResponse:
-        team_members: List[UserEntity] = []
-        for user in project_data.team_members:
-            user_entity = self.db.query(UserEntity).filter(UserEntity.email == user.email).first()
-            team_members.append(user_entity)        
-
-        project_leaders: List[UserEntity] = []
-        for leader in project_data.project_leaders:
-            leaders_entity = self.db.query(UserEntity).filter(UserEntity.email == leader.email).first()
-            project_leaders.append(leaders_entity)
+        team_members = self._get_user_entities_by_emails(project_data.team_members)
+        project_leaders = self._get_user_entities_by_emails(project_data.project_leaders)
 
         new_project_entity = ProjectEntity.from_model(project_data, team_members, project_leaders)
         self.db.add(new_project_entity)
@@ -33,7 +25,7 @@ class ProjectService:
 
         return new_project_entity.to_project_response()
 
-    def update_project(self, project_id: int, project_update: ProjectUpdate) -> ProjectResponse:
+    def update_project(self, project_id: int, project_update: ProjectUpdate, current_user_id: int) -> ProjectResponse:
         logger.info(f"Starting update for project with id {project_id}")
 
         try:
@@ -42,6 +34,11 @@ class ProjectService:
         except NoResultFound:
             logger.error(f"Project with id {project_id} not found")
             raise ProjectNotFoundException(f"Project with id {project_id} not found")
+
+        # Check if the current user is a project leader
+        if current_user_id not in [leader.id for leader in project_entity.project_leaders]:
+            logger.error(f"User with id {current_user_id} is not authorized to update project with id {project_id}")
+            raise UnauthorizedException(f"User with id {current_user_id} is not authorized to update project with id {project_id}")
 
         update_data = project_update.model_dump(exclude_unset=True)
 
@@ -88,6 +85,11 @@ class ProjectService:
         for user in users_to_add:
             project_entity.project_leaders.append(user)
 
+        # Ensure project leaders are not empty
+        if not project_entity.project_leaders:
+            logger.error(f"Cannot update project with id {project_id} because it would leave the project without leaders")
+            raise ValueError("Project must have at least one leader")
+
         # Update other fields
         for field, value in update_data.items():
             if field not in ["team_members", "project_leaders"]:
@@ -100,7 +102,6 @@ class ProjectService:
         logger.info(f"Project with id {project_id} refreshed from the database")
 
         return project_entity.to_project_response()
-
 
     def get_project(self, project_id: int) -> ProjectResponse:
         project = self.db.query(ProjectEntity).filter_by(id=project_id).first()
@@ -116,11 +117,31 @@ class ProjectService:
         projects = self.db.query(ProjectEntity).filter(ProjectEntity.project_leaders.any(UserEntity.id == user_id)).all()
         return [project.to_project_response() for project in projects]
 
-    def delete_project(self, project_id: int):
+    def delete_project(self, project_id: int, current_user_id: int) -> ProjectResponse:
         project_entity = self.db.query(ProjectEntity).filter_by(id=project_id).first()
         if project_entity is None:
             raise ProjectNotFoundException(f"Project with id {project_id} not found")
+        
+        # Check if the current user is one of the project leaders
+        if current_user_id not in [leader.id for leader in project_entity.project_leaders]:
+            raise UnauthorizedException("You do not have permission to delete this project")
+        
         self.db.delete(project_entity)
         self.db.commit()
         return project_entity.to_project_response()
 
+    def leave_project(self, project_id: int, current_user_id: int) -> ProjectResponse:
+        project_entity = self.db.query(ProjectEntity).filter_by(id=project_id).first()
+        if project_entity is None:
+            raise ProjectNotFoundException(f"Project with id {project_id} not found")
+        
+        # Check if the current user is one of the team members
+        if current_user_id not in [member.id for member in project_entity.team_members]:
+            raise UnauthorizedException("You are not a member of this project")
+        
+        project_entity.team_members = [member for member in project_entity.team_members if member.id != current_user_id]
+        self.db.commit()
+        return project_entity.to_project_response()
+
+    def _get_user_entities_by_emails(self, users: List[UserResponse]) -> List[UserEntity]:
+        return [self.db.query(UserEntity).filter(UserEntity.email == user.email).first() for user in users]
